220208_강주수_몬스터애니

+++ Astar_Corp_ThreeTeam/Assets/Scripts/Kang/Battle/BattleTile.cs
@@ -12,7 +12,7 @@ public class BattleTile : MonoBehaviour
     {
         var dics = BattleMgr.Instance.tileMgr.tileDics;
         var pos = transform.position;
-        tileIdx = new Vector3(pos.x, pos.y - 0.5f, pos.z);
+        tileIdx = new Vector3(pos.x, pos.y - 0.5f, (int)pos.z);
         foreach (var pair in dics)
         {
             var tile = pair.Value;
